requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
pyyaml>=6.0.1
httpx>=0.26.0
sse-starlette>=2.0.0
